ref(InviteModal): Extract some render-related function to their own react files
